Made the test incorrect
made the test incorrect

+++ src/test/java/CalculatorTest.java
@@ -6,7 +6,7 @@ public class CalculatorTest {
 
     @Test
     public void testNullInputs() {
-        assertEquals(new Integer(1),calculator.sumAllNumbers(0,1).get());
+        assertEquals(new Integer(5),calculator.sumAllNumbers(0,1).get());
     }
     
 }
